Rename output to azure_monitor along with README updates
